Implementacion de transaccion Matricular

+++ Entities/Alumno.cs
@@ -6,7 +6,7 @@ namespace Entities
 {
         public class Alumno
         {
-            public int IdAlumno { get; set; }
+            public long IdAlumno { get; set; }
             public string Nombre { get; set; }
             public string Apellido { get; set; }
             public string Dni { get; set; }
